TS => MJS: types/range_map
Migrate fjs/types/range_map/module.f.ts to module.f.mjs under the
stage-1 TypeScript-to-mjs migration, folded into the
patricia_trie/sorted_list PR since it became a fresh dependency leaf
once sorted_list migrated (all of range_map's other runtime
dependencies -- list, nullable, number, function/compare,
function/operator, range -- were already .f.mjs).

proof.f.ts stays TypeScript for now: it still imports the unmigrated
media/json, object, sorted_set, and string modules; only the module
import specifier is updated.

Same per-arrow @template/@param JSDoc style as sorted_list for the
curried, generic reduceOp/tailReduce/merge/get/fromRange/rangeMap
helpers, for the same nested-generic-inference reason.

Update the six remaining .f.ts importers (range_set, byte_set,
bnf/ll1, fsm, fsc, js/tokenizer) and three todo docs that quoted the
old .f.ts path or line numbers.
