+++ fjs/fsm/todo/scan-as-map.md
@@ -39,9 +39,9 @@ const scanFetch = map(([item]: Entry<SortedSet<string>>) => item)
 ```
 
 Note the explicit `Entry<string>` return annotation instead of `as const`:
-`Entry<T>` (`fjs/types/range_map/module.f.ts:46`) is the *mutable* tuple
+`Entry<T>` (`fjs/types/range_map/module.f.mjs:48`) is the *mutable* tuple
 `[T, number]`, and `Dfa` stores `RangeMapArray<string> = readonly
-Entry<string>[]` (`:54`), so a `readonly [string, number]` produced by
+Entry<string>[]` (`:58`), so a `readonly [string, number]` produced by
 `as const` would not be assignable where
 `toArray(scanStringify(setMap))` lands in `newDfa`. The annotation pins the
 tuple type without widening and matches what `stringifyOp` returns today.
