+++ fjs/js/todo/174.md
@@ -95,4 +95,4 @@ its `CreateToToken` continuation directly over `NumberRange`.
 ### Related
 
 - [i165](../../issues/README.md) — layered BNF parser (different mechanism).
-- `fjs/types/range_map/module.f.ts` — the underlying structure both consumers wrap.
+- `fjs/types/range_map/module.f.mjs` — the underlying structure both consumers wrap.
