+++ fjs/types/byte_set/module.f.ts
@@ -4,7 +4,7 @@
  * @module
  */
 import { compose } from '../function/module.f.mjs'
-import type { RangeMap } from '../range_map/module.f.ts'
+import type { RangeMap } from '../range_map/module.f.mjs'
 import type { SortedSet } from '../sorted_set/module.f.ts'
 import { reverse, countdown, flat, map } from '../list/module.f.mjs'
